- Removed subroutines because they seem to be non-portable enough. Also, no one likes them. - Moved towards a bit more direct control when it comes to command buffer and per-draw uniform variables constant buffer sizes.

+++ include/tempest/utils/patterns.hh
@@ -106,7 +106,7 @@ struct PackedData
 };
 
 template<class T, class... TArgs>
-T* CreatePackedData(size_t count, TArgs&&... args)
+T* CreatePackedData(uint32 count, TArgs&&... args)
 {
     return new (malloc(sizeof(T) + count*sizeof(typename T::PackType))) T(count, args...);
 }
@@ -120,7 +120,7 @@ void DestroyPackedData(T* ptr)
 
 #define PACKED_DATA(type) \
     typedef type PackType; \
-    template<class T, class... TArgs> friend T* Tempest::CreatePackedData(size_t count, TArgs&&... args); \
+    template<class T, class... TArgs> friend T* Tempest::CreatePackedData(uint32 count, TArgs&&... args); \
     template<class T> friend void Tempest::DestroyPackedData(T* ptr); \
     Tempest::PackedData<type>
